requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
